+++ cypress/integration/views.js
@@ -91,4 +91,46 @@ describe("View ID tests ", () => {
             });
         });
     });
+
+    // ===========================
+    //  Confirms:
+    // view A's id and event A's cvid are same
+    // view B's id and event B's cvid are same. Also view B's pvid and view A's id are same
+    // view C's id and event C's cvid are same. Also view C's pvid and view B's id are same
+    //
+    // request order: view A start -> event A -> view A end -> view B start -> event B -> view B end -> view C start -> event C  
+    // ===========================
+    it("Checks a sequence of events and page views", () => {
+        hp.haltAndClearStorage(() => {
+            initMain();
+            Countly.track_view("A");
+            Countly.add_event({ key: "A" });
+            Countly.track_view("B");
+            Countly.add_event({ key: "B" });
+            Countly.track_view("C");
+            Countly.add_event({ key: "C" });
+
+            cy.fetch_local_event_queue().then((eq) => {
+                expect(eq.length).to.equal(8);
+                cy.log(eq);
+
+                // event A and view A
+                cy.check_view_event(eq[0], "A", undefined, false); // no pvid
+                const idA = eq[0].id; // idA
+                cy.check_event(eq[1], { key: "A" }, undefined, idA); // cvid should be idA
+                cy.check_view_event(eq[2], "A", 0, false); // no pvid
+
+                // event B and view B
+                cy.check_view_event(eq[3], "B", undefined, idA); // pvid is idA
+                const idB = eq[3].id; // idB
+                cy.check_event(eq[4], { key: "B" }, undefined, idB); // cvid should be idB
+                cy.check_view_event(eq[5], "B", 0, idA); // pvid is idA
+
+                // event C and view C
+                cy.check_view_event(eq[6], "C", undefined, idB); // pvid is idB
+                const idC = eq[6].id; // idC
+                cy.check_event(eq[7], { key: "C" }, undefined, idC); // cvid should be idC  
+            });
+        });
+    });
 });