+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 23.02.0
 - Events are now recorded with an internal ID.
 - Mitigated an issue where users could have truncate an internal event key
+- Mitigated an issue where SDK was reporting empty scroll values
 
 ## 22.06.5
 - SDK now adds userAgent string to each request to prevent proxy related issues